Make <Leader>zk handle a custom opfunc that uses the visual selection itself
KeepText#InSelection#KeepText() assumed that by the time it's invoked (after the g@), the original selection done by the user is still intact. That's correct for built-in motions and text objects. Custom text objects typically use the visual selection to indicate the area they apply to. That clashes with the use of the selection as a mean to mark the original area here.
We need to save the selection earlier; i.e. before the g@ triggers the opfunc. A new KeepText#InSelection#SaveSelection() wrapper used in <Plug>(KeepTextSelectionOperator) provides a way to do this.
The local variables are changed to script-scoped s:selectionStartPos, s:selectionEndPos, s:selectionStartVirtCol, s:selectionEndVirtCol, s:selectionMode.
KeepText#InSelection#KeepText() also uses gv + yank to obtain the selected text. We need to restore the original selection, but do so only if actually necessary, to avoid needless operations (and potential flickering) when a built-in motion or text object was used.

System-info: Ubuntu 20.04.6 LTS, x86_64
Platform-info: Vim 9.1.1029